feat: Çift CTRL ile 2D/3D geçişi
- ui.js: GSAP animasyonlu toggle3DPerspective eklendi
- input.js: Çift CTRL tuş algılaması eklendi (300ms)
- pointer-up.js: Eski Ctrl + Orta Tuş kodu kaldırıldı

Artık kullanıcılar CTRL tuşuna hızlıca iki kez basarak
2D ve 3D perspektif arasında yumuşak geçiş yapabilir.
